requirements.txt
appdirs==1.4.3
Django==1.8.3
django-extensions==1.7.4
django-twitter-bootstrap==3.3.0
mysql==0.0.1
MySQL-python==1.2.5
packaging==16.8
pygeoip==0.3.2
pyparsing==2.2.0
six==1.10.0
virtualenv==15.0.3
mysqlclient==1.3.6
python-dateutil==2.6.0
